Version 4.1
- Automatically define iabbreviations by adding [bg]:tskelAbbrevPostfix
(default: '#') to the bit name (i.e., a bit with the file "foo.bar" will
by default create the menu entry "TSkel.foo.bar" for the bit "bar" and
the abbreviation "bar#"). If this causes problems, set
g:tskelAutoAbbrevs to 0.
- Bits can have a <tskel:abbrev> section that defines the abbreviation.
- New type 'abbreviations': This will make your abbreviations accessible
as templates (in case you can't remember their names)
- New experimental <tskel:condition> section (a vim expression) that
checks if a bit is eligible in the current context.
- New <+input()+> tag.
- New <+execute()+> tag.
- New <+let(VAR=VALUE)+> tag.
- <+include(NAME)+> as synonym for <+bit:NAME+>.
- Experimental <+if()+> ... <+elseif()+> ... <+else+> ... <+endif+>,
<+for(var in list)+> ... <+endfor+> tags.
- Special tags <+nop+>, <+joinline+>, <+nl+> to prevent certain
problems.
- These special tags have to be lower case.
- Made tskeleton#GoToNextTag() smarter in recognizing something like:
<+/DEFAULT+>.
- Defined <Leader>## and <Leader>#<space> (see g:tskelMapLeader) as
visual command (the user will be queried for the name of a skeleton)
- Some functions have moved and changed names. It should now be possible
to plug-in custom template expanders (or re-use others).
- Use append() via tlib#buffer#InsertText() to insert bits. This could
cause old problems to reappear although it seems to work fine.
- The markup should now be properly configurable (per buffer; you can
set template-specific markers in the tskel:here_before section).
- Require tlib 0.14
- The default value for g:tskelUseBufferCache is 0 as many people might
find the accumulation of cached information somewhat surprising. Unless
you use tag/functions type of skeleton bit, it's unnecessary anyway.
- Removed the dependency on genutils.
- The g:tskelMarkerCursor variable was removed and replaced with
g:tskelMarkerCursor_mark and g:tskelMarkerCursor_rx.